+++ examples/c++/qt/deep-diver-with-c++-api/qtsail.h
@@ -23,6 +23,8 @@
 #ifndef QT_SAIL_H
 #define QT_SAIL_H
 
+#include <cstddef>
+
 #include <QWidget>
 #include <QScopedPointer>
 
@@ -51,8 +53,8 @@ public:
 private:
     sail_error_t init();
     sail_error_t loadImage(const QString &path, QImage *qimage);
-    sail_error_t saveImage(const QImage &qimage, void *buffer, unsigned long buffer_length,
-                           unsigned long *written);
+    sail_error_t saveImage(const QImage &qimage, void *buffer, size_t buffer_length,
+                           size_t *written);
     sail_error_t pluginInfo(const sail::plugin_info &plugin_info) const;
     QStringList filters() const;
