+++ src/libsail/src/io_mem.h
@@ -19,6 +19,8 @@
 #ifndef SAIL_IO_MEM_H
 #define SAIL_IO_MEM_H
 
+#include <stddef.h>
+
 #ifdef SAIL_BUILD
     #include "error.h"
     #include "export.h"
@@ -35,7 +37,7 @@ struct sail_io;
  *
  * Returns 0 on success or sail_error_t on error.
  */
-SAIL_HIDDEN sail_error_t alloc_io_read_mem(const void *buffer, unsigned long buffer_length, struct sail_io **io);
+SAIL_HIDDEN sail_error_t alloc_io_read_mem(const void *buffer, size_t buffer_length, struct sail_io **io);
 
 /*
  * Opens the specified memory buffer for writing and allocates a new I/O object for it.
@@ -43,6 +45,6 @@ SAIL_HIDDEN sail_error_t alloc_io_read_mem(const void *buffer, unsigned long buf
  *
  * Returns 0 on success or sail_error_t on error.
  */
-SAIL_HIDDEN sail_error_t alloc_io_write_mem(void *buffer, unsigned long buffer_length, struct sail_io **io);
+SAIL_HIDDEN sail_error_t alloc_io_write_mem(void *buffer, size_t buffer_length, struct sail_io **io);
 
 #endif
