+++ src/libsail/src/io_noop.h
@@ -20,7 +20,7 @@
 #define SAIL_IO_NOOP_H
 
 #include <stdbool.h>
-#include <stdio.h>
+#include <stddef.h>
 
 #ifdef SAIL_BUILD
     #include "error.h"
@@ -35,7 +35,7 @@
  */
 SAIL_HIDDEN sail_error_t io_noop_read(void *stream, void *buf, size_t object_size, size_t objects_count, size_t *read_objects_count);
 SAIL_HIDDEN sail_error_t io_noop_seek(void *stream, long offset, int whence);
-SAIL_HIDDEN sail_error_t io_noop_tell(void *stream, unsigned long *offset);
+SAIL_HIDDEN sail_error_t io_noop_tell(void *stream, size_t *offset);
 SAIL_HIDDEN sail_error_t io_noop_write(void *stream, const void *buf, size_t object_size, size_t objects_count, size_t *written_objects_count);
 SAIL_HIDDEN sail_error_t io_noop_flush(void *stream);
 SAIL_HIDDEN sail_error_t io_noop_close(void *stream);
