+++ src/libsail/src/sail_deep_diver.h
@@ -19,6 +19,8 @@
 #ifndef SAIL_SAIL_DEEP_DIVER_H
 #define SAIL_SAIL_DEEP_DIVER_H
 
+#include <stddef.h>
+
 #ifdef SAIL_BUILD
     #include "error.h"
     #include "export.h"
@@ -82,7 +84,7 @@ SAIL_EXPORT sail_error_t sail_start_reading_file_with_options(const char *path, 
  *
  * Returns 0 on success or sail_error_t on error.
  */
-SAIL_EXPORT sail_error_t sail_start_reading_mem_with_options(const void *buffer, unsigned long buffer_length, struct sail_context *context,
+SAIL_EXPORT sail_error_t sail_start_reading_mem_with_options(const void *buffer, size_t buffer_length, struct sail_context *context,
                                                              const struct sail_plugin_info *plugin_info,
                                                              const struct sail_read_options *read_options, void **state);
 
@@ -131,7 +133,7 @@ SAIL_EXPORT sail_error_t sail_start_writing_file_with_options(const char *path, 
  *
  * Returns 0 on success or sail_error_t on error.
  */
-SAIL_EXPORT sail_error_t sail_start_writing_mem_with_options(void *buffer, unsigned long buffer_length, struct sail_context *context,
+SAIL_EXPORT sail_error_t sail_start_writing_mem_with_options(void *buffer, size_t buffer_length, struct sail_context *context,
                                                              const struct sail_plugin_info *plugin_info,
                                                              const struct sail_write_options *write_options, void **state);
 
@@ -144,7 +146,7 @@ SAIL_EXPORT sail_error_t sail_start_writing_mem_with_options(void *buffer, unsig
  *
  * Returns 0 on success or sail_error_t on error.
  */
-SAIL_EXPORT sail_error_t sail_stop_writing_with_written(void *state, unsigned long *written);
+SAIL_EXPORT sail_error_t sail_stop_writing_with_written(void *state, size_t *written);
 
 /* extern "C" */
 #ifdef __cplusplus
